Revert "Merge branch 'torsiondrive' into xtb"
This reverts commit 0b81cc4ba561e7e4a0f4ae417c6fdc5e90f3bc13, reversing
changes made to 05926698f4b30946164e2a080e914ab7bc23fc8d.
