+++ qforce/qm/qm.py
@@ -1,17 +1,14 @@
 import os
-import shutil
 import sys
 from ase.io import read, write
 import numpy as np
 from colt import Colt
-from warnings import warn
 #
 from .gaussian import Gaussian
 from .qchem import QChem
 from .orca import Orca
-from .xtb import xTB
-from .qm_base import scriptify, HessianOutput, ScanOutput, ReadABC
-from ..elements import ATOM_SYM
+from .xtb import xTB 
+from .qm_base import scriptify, HessianOutput, ScanOutput
 
 
 implemented_qm_software = {'gaussian': Gaussian,
@@ -45,10 +42,6 @@ n_proc = 1 :: int
 
 # Scaling of the vibrational frequency for the corresponding QM method (not implemented)
 vib_scaling = 1.0 :: float
-
-# Use the internal relaxed scan method of the QM software or the Torsiondrive method using xTB
-dihedral_scanner = relaxed_scan :: str :: [relaxed_scan, xtb-torsiondrive]
-
 """
     _method = ['scan_step_size']
 
@@ -68,130 +61,10 @@ dihedral_scanner = relaxed_scan :: str :: [relaxed_scan, xtb-torsiondrive]
         n_scan_steps = int(np.ceil(360/self.config.scan_step_size))
 
         for file in files:
-            if self.config.dihedral_scanner == 'relaxed_scan':
-                qm_outs.append(self.software.read().scan(self.config, f'{self.job.frag_dir}/{file}'))
-            elif self.config.dihedral_scanner == 'xtb-torsiondrive':
-                qm_outs.append(self._xtb_torsiondrive_read(self.config, f'{self.job.frag_dir}/{file}'))
+            qm_outs.append(self.software.read().scan(self.config, f'{self.job.frag_dir}/{file}'))
         qm_out = self._get_unique_scan_points(qm_outs, n_scan_steps)
 
         return ScanOutput(file, n_scan_steps, *qm_out)
-
-    def _xtb_torsiondrive_read(self, config, log_file):
-        '''Read the TorsionDrive output.
-
-        Parameters
-        ----------
-        config : config
-            A configparser object with all the parameters.
-        log_file : string
-            The path to the fragment result output.
-
-        Returns
-        -------
-        n_atoms : int
-            The number of atoms in the molecule.
-        coords : list
-            A list of array of float. The list has the length of the number
-            of steps and the array has the shape of (n_atoms, 3).
-        angles : list
-            A list (length: steps) of the angles that is being scanned.
-        energies : list
-            A list (length: steps) of the energy.
-        point_charges : dict
-            A dictionary with key in charge_method and the value to be a
-            list of float of the size of n_atoms.
-        '''
-        with open(log_file, 'r') as f:
-            coord_text = f.read()
-        lines = coord_text.strip().split('\n')
-        n_atoms = int(lines[0])
-        num_conf = len(lines) // (n_atoms+2)
-        energy_list = []
-        coord_list = []
-        angle_list = []
-        for conf in range(num_conf):
-            current = lines[conf*(n_atoms+2):(conf+1)*(n_atoms+2)]
-            _, angle, _, energy = current[1].split()
-            angle = float(angle[1:-2])
-            angle_list.append(angle)
-            energy = float(energy)
-            energy_list.append(energy)
-            n_atoms, elements, coords = ReadABC._read_xyz_coords('\n'.join(current))
-            coord_list.append(coords)
-
-        point_charges = np.loadtxt(os.path.splitext(log_file)[0]+'.charges')
-        return n_atoms, coord_list, angle_list, energy_list, {'xtb': point_charges}
-
-    def _xtb_torsiondrive_write(self, file, dir, scan_id, coords, atnums,
-                                   scanned_atoms, charge,
-                                   multiplicity):
-        '''Create the TorsionDrive input.
-
-        Parameters
-        ----------
-        file : file
-            The file object to write the torsiondrive command
-        dir : string
-            The path to the current directory.
-        scan_id : string
-            The name of the fragment.
-        coords : numpy.array
-            Array of the coordinates in the same of (n,3), where n is the
-            number of atoms.
-        atnums : list
-            A list of n integers representing the atomic number, where n is
-            the number of atoms.
-        scanned_atoms : list
-            A list of 4 integers representing the one-based atom index of the
-            dihedral.
-        charge : int
-            The total charge of the fragment.
-        multiplicity : int
-            The multiplicity of the molecule.
-         '''
-        try:
-            os.makedirs(f'{dir}/{scan_id}_torsiondrive')
-        except FileExistsError:
-            warn(f'Folder {dir}/{scan_id}_torsiondrive exists, remove the folder.')
-            shutil.rmtree(f'{dir}/{scan_id}_torsiondrive')
-            os.makedirs(f'{dir}/{scan_id}_torsiondrive')
-
-        # Create the coordinate input file
-        with open(f'{dir}/{scan_id}_torsiondrive/input.xyz', 'w') as f:
-            n_atoms = len(atnums)
-            uhf = multiplicity - 1
-            cmd = f'xTB arguments: --opt --chrg {charge} --uhf {uhf} --gfn 2 --parallel 1'
-            crd = '\n'.join(['{:>3s} {:>12.6f} {:>12.6f} {:>12.6f}'.format(ATOM_SYM[ele_id], *pos)
-                             for ele_id, pos in zip(atnums, coords)])
-            f.write(f'''{n_atoms}
-{cmd}
-{crd}''')
-        with open(f"{dir}/{scan_id}_torsiondrive/dihedrals.txt", 'w') as f:
-            f.write('{} {} {} {}'.format(*scanned_atoms))
-
-        # Write the torsiondrive input command
-        file.write(f'cd {scan_id}_torsiondrive\n')
-        file.write(f'torsiondrive-launch input.xyz dihedrals.txt '
-                   f'-g {int(self.config.scan_step_size)} '
-                   f'-e xtb --native_opt -v \n')
-        file.write(f'cp scan.xyz ../{scan_id}.log\n')
-        file.write(f'cp opt_tmp/gid_+000/1/charges ../{scan_id}.charges\n')
-        file.write(f'cp opt_tmp/gid_+180/1/charges ../{scan_id}.charges\n')
-        file.write(f'cd ..\n')
-
-        # cwd = os.getcwd()
-        # os.chdir(f"{dir}/{scan_id}_torsiondrive")
-        # grid_ids, grid_energies, grid_geometries, elements = run("input.xyz",
-        #     [self.config.scan_step_size, ], 'dihedrals.txt', verbose=False)
-        # os.chdir(cwd)
-        #
-        # coord_list = []
-        # energy_list = []
-        # for grid_id in grid_ids:
-        #     coord_list.append(grid_geometries[grid_id])
-        #     energy_list.append(grid_energies[grid_id])
-        #
-        # return grid_ids, coord_list, energy_list
 
     @scriptify
     def write_hessian(self, file, coords, atnums):
@@ -200,39 +73,8 @@ dihedral_scanner = relaxed_scan :: str :: [relaxed_scan, xtb-torsiondrive]
     @scriptify
     def write_scan(self, file, scan_id, coords, atnums, scanned_atoms, start_angle, charge,
                    multiplicity):
-        '''Generate the input file for the dihedral scan.
-
-        Parameters
-        ----------
-        file : file
-            The file handler for writing the input file. e.g. file.write('test')
-        scan_id : string
-            The name of the fragment.
-        coords : numpy.array
-            Array of the coordinates in the same of (n,3), where n is the
-            number of atoms.
-        atnums : list
-            A list of n integers representing the atomic number, where n is
-            the number of atoms.
-        scanned_atoms : list
-            A list of 4 integers representing the one-based atom index of the
-            dihedral.
-        start_angle : float
-            The dihedral angle of the current conformation.
-        charge : int
-            The total charge of the fragment.
-        multiplicity : int
-            The multiplicity of the molecule.
-        '''
-        if self.config.dihedral_scanner == 'relaxed_scan':
-            self.software.write().scan(file, scan_id, self.config, coords,
-                                       atnums, scanned_atoms, start_angle,
-                                       charge, multiplicity)
-        elif self.config.dihedral_scanner == 'xtb-torsiondrive':
-            self._xtb_torsiondrive_write(file, self.job.frag_dir, scan_id, coords,
-                                            atnums, scanned_atoms, charge,
-                                            multiplicity)
-
+        self.software.write().scan(file, scan_id, self.config, coords, atnums, scanned_atoms,
+                                   start_angle, charge, multiplicity)
 
     def _get_unique_scan_points(self, qm_outs, n_scan_steps):
         all_angles, all_energies, all_coords, chosen_point_charges, final_e = [], [], [], {}, 0
